Add mariadb-upgrade on MARIADB_AUTO_UPGRADE=1
Per docs[1], this attempts to backup the system mysql database
if there is a major version change. If we can't find the previous
version (from mysql_upgrade_info), back it up anyway.

Backups are stored on the top level of the datadir as a mariadb-dump
of the mysql database, zstd compressed.

Both the backup and the mariadb-upgrade require MARIADB_ROOT_PASSWORD
to be set.

mariadb-upgrade is run on the final running instance. The rational is:
* There should be no impacts of system tables.
* The default check step of mariadb-upgrade could be quite time
  consuming as it iterates over all the user databases/tables.
* The iteration is a set of mariadb-check on all the tables. There is
  a bit of IO here, one at a time, and sequential. From 10.2 onwards
  even the analyze table is read view and non-blocking.
* If users encounter delays on this step, they can add to their config
  file to skip there user databases:
  [mariadb-upgrade]
  upgrade-system-tables

Running mariadb-upgrade on the datadir after initalization was timing
out test cases. Maybe we can add it to mysql_install_db upstream.

Currently the default is to not to upgrade. At some point the
MARIADB_AUTO_UPGRADE may default to on (1).

[1] https://mariadb.com/kb/en/mysql_upgrade/

Closes #350